integrate shared header files into the user test files
