+++ src/Makefile.test_util.include
@@ -19,8 +19,6 @@ TEST_UTIL_H = \
   blsct/public_key.h \
   blsct/public_keys.h \
   blsct/signature.h \
-  test/blsct/arith/shared_point_tests.h \
-  test/blsct/arith/shared_scalar_tests.h \
   test/util/blockfilter.h \
   test/util/chainstate.h \
   test/util/coins.h \
